[LFC][Floating] Containing block of a float could push the candidate position beyond the current float. https://bugs.webkit.org/show_bug.cgi?id=188264
Reviewed by Antti Koivisto.

<div id=first style="float: left; width: 100px; height: 10px;"></div>
<div id=second style="float: left; width: 10px; height: 10px;"></div>
<div style="width: 50px; height: 50px;>
    <div id=third style="float: left; width: 10px; height: 10px;"></div>
</div>
In the example above by the time the "third" float comes in, we are already at 110px horizontally. However the containing block of the "third" float forces
the candidate position at 0px. In such cases intersecting won't work since the candidate position is beyond the last(second) float (horizontally).
We could either keep going back to the previous floats on this vertical position and eventually intersect with "first" or just compare the horizontal
positions and ensure that the left/right float is not going to be placed to the left/right of the last float on the current vertical position.

* layout/FloatingContext.cpp:
(WebCore::Layout::FloatingPair::intersects const):

Identifier: 203375@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234508 268f45cc-cd09-0410-ab3c-d52691b4dbfc