[iOS] Indeterminate checkboxes appear unchecked https://bugs.webkit.org/show_bug.cgi?id=160484 <rdar://problem/31143185>
Patch by Aditya Keerthi <akeerthi@apple.com> on 2018-08-06
Reviewed by Wenson Hsieh.

Source/WebCore:

Add support for rendering indeterminate checkboxes on iOS.

We now paint the unchecked appearance only if the checkbox is not checked and is
not indeterminate. Otherwise, we draw a checkmark if the checkbox is in the
checked state and a dash if the checkbox is in the indeterminate state. Both of
these states share the same dark background.

Test: fast/forms/ios/render-indeterminate-checkbox.html

* css/html.css:
(input[type="checkbox"]:indeterminate):
(input[type="checkbox"]:indeterminate:disabled):
* rendering/RenderThemeIOS.mm:
(WebCore::drawJoinedLines):
(WebCore::RenderThemeIOS::paintCheckboxDecorations):

LayoutTests:

Added reftest to ensure that the indeterminate checkbox does not appear the same
as an unchecked checkbox.

* fast/forms/ios/render-indeterminate-checkbox-expected-mismatch.html: Added.
* fast/forms/ios/render-indeterminate-checkbox.html: Added.
* platform/ios/fast/forms/indeterminate-expected.txt: Rebaseline.

Identifier: 203454@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234609 268f45cc-cd09-0410-ab3c-d52691b4dbfc