[JSC] Should not rotate constant with 64 https://bugs.webkit.org/show_bug.cgi?id=188556
Reviewed by Mark Lam.

To defend against JIT splaying, we rotate a constant with a randomly generated seed.
But if a seed becomes 64, the following code performs `value << 64` where value's type
is uint64_t, and it causes undefined behaviors (UBs). This patch limits the seed in the
range of [0, 64) not to generate code causing UBs. This is found by UBSan.

* assembler/MacroAssembler.h:
(JSC::MacroAssembler::generateRotationSeed):
(JSC::MacroAssembler::rotationBlindConstant):

Identifier: 203647@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@234852 268f45cc-cd09-0410-ab3c-d52691b4dbfc